requirements.txt
fastapi~=0.111.0
uvicorn~=0.29.0
openai==0.28
python-dotenv~=1.0.1

pydantic~=2.7.1